+++ ocrd/ocrd/decorators.py
@@ -56,25 +56,21 @@ def ocrd_cli_wrap_processor(
             raise Exception(msg)
         resolver = Resolver()
         workspace = resolver.workspace_from_url(mets, working_dir)
+        page_id = kwargs.get('page_id')
         if overwrite:
             if 'output_file_grp' not in kwargs or not kwargs['output_file_grp']:
                 raise Exception("--overwrite requires --output-file-grp")
             LOG.debug("Removing files because of --overwrite")
             for grp in kwargs['output_file_grp'].split(','):
-                if 'page_id' in kwargs:
-                    for page_id in kwargs['page_id'].split(','):
-                        LOG.debug("Removing files in output file group %s with page ID %s", grp, page_id)
-                        for file in workspace.mets.find_files(pageId=page_id, fileGrp=grp):
+                if page_id:
+                    for one_page_id in kwargs['page_id'].split(','):
+                        LOG.debug("Removing files in output file group %s with page ID %s", grp, one_page_id)
+                        for file in workspace.mets.find_files(pageId=one_page_id, fileGrp=grp):
                             workspace.remove_file(file, force=True, keep_file=False, page_recursive=True)
                 else:
                     LOG.debug("Removing all files in output file group %s ", grp)
                     workspace.remove_file_group(grp, recursive=True, force=True, keep_files=False, page_recursive=True, page_same_group=True)
-        if overwrite and 'page_id' in kwargs:
-            # If 'overwrite' and 'page_id' were given, it is possible the output
-            # file group is not completely empty
-            report = WorkspaceValidator.check_file_grp(workspace, kwargs['input_file_grp'], '')
-        else:
-            report = WorkspaceValidator.check_file_grp(workspace, kwargs['input_file_grp'], kwargs['output_file_grp'])
+        report = WorkspaceValidator.check_file_grp(workspace, kwargs['input_file_grp'], kwargs['output_file_grp'], page_id)
         if not report.is_valid:
             raise Exception("Invalid input/output file grps:\n\t%s" % '\n\t'.join(report.errors))
         run_processor(processorClass, ocrd_tool, mets, workspace=workspace, **kwargs)